+++ internal/transport/transport.go
@@ -3,40 +3,40 @@ package transport
 
 import (
 	"fyne.io/fyne/v2"
-	"github.com/psanford/wormhole-william/wormhole"
+	"github.com/rymdport/wormhole/wormhole"
 )
 
 // Client defines the client for handling sending and receiving using wormhole-william
 type Client struct {
 	wormhole.Client
 
-	app fyne.App
+	// App is a reference to the currently running Fyne application.
+	App fyne.App
 
-	// Save a reference to the window to avoid creating a new one when sending and receiving text.
-	textSendWindow *textSendWindow
-	textRecvWindow *textRecvWindow
+	// CustomCode defines if we should pass a custom code or let wormhole-william generate on for us.
+	CustomCode bool
+
+	// DownloadPath holds the download path used for saving received files.
+	DownloadPath string
+
+	// NoExtractDirectory specifies if we should extract directories or not.
+	NoExtractDirectory bool
 
 	// Notification holds the settings value for if we have notifications enabled or not.
 	Notifications bool
 
 	// OverwriteExisting holds the settings value for if we should overwrite already existing files.
 	OverwriteExisting bool
-
-	// DownloadPath holds the download path used for saving received files.
-	DownloadPath string
-
-	// Defines if we should pass a custom code or let wormhole-william generate on for us.
-	CustomCode bool
 }
 
 // ShowNotification sends a notification if c.Notifications is true.
 func (c *Client) ShowNotification(title, content string) {
 	if c.Notifications {
-		c.app.SendNotification(&fyne.Notification{Title: title, Content: content})
+		c.App.SendNotification(&fyne.Notification{Title: title, Content: content})
 	}
 }
 
 // NewClient returns a new client for sending and receiving using wormhole-william
 func NewClient(app fyne.App) *Client {
-	return &Client{app: app}
+	return &Client{App: app}
 }